+++ tests/test_bench.py
@@ -224,6 +224,22 @@ def test_graphstate_seed() -> None:
     assert qc_no_seed.name == "graphstate"
 
 
+def test_dynamical_qft() -> None:
+    """Test the creation of the Dynamical Quantum Fourier Transform + Measurement."""
+    qc = create_circuit("dynamical_qft", 5)
+    assert qc.num_qubits == 5
+    assert qc.num_clbits == 5
+    assert qc.name == "dynamical_qft"
+
+    # Check the number of Hadamard gates, Measurements and if_else blocks
+    hadamard_count = sum(1 for inst in qc.data if inst.operation.name == "h")
+    measurement_count = sum(1 for inst in qc.data if inst.operation.name == "measure")
+    if_else_count = sum(1 for inst in qc.data if inst.operation.name == "if_else")
+    assert hadamard_count == 5
+    assert measurement_count == 5
+    assert if_else_count == 5
+
+
 @pytest.mark.parametrize(
     (
         "benchmark_name",