+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 
 - ✨ Add support for compiling `if_test` for all gatesets and devices ([#805]) ([**@adnathanail**])
+- ✨ Add Dynamical QFT + Measurement circuit as a dynamical benchmarking circuit ([#807]) ([**@morrishuismans**])
 
 ### Changed
 
@@ -97,6 +98,7 @@ _📚 Refer to the [GitHub Release Notes] for previous changelogs._
 
 <!-- PR links -->
 
+[#807]: https://github.com/munich-quantum-toolkit/bench/pull/807
 [#805]: https://github.com/munich-quantum-toolkit/bench/pull/805
 [#794]: https://github.com/munich-quantum-toolkit/bench/pull/794
 [#731]: https://github.com/munich-quantum-toolkit/bench/pull/731
@@ -144,6 +146,7 @@ _📚 Refer to the [GitHub Release Notes] for previous changelogs._
 [**@denialhaag**]: https://github.com/denialhaag
 [**@soroushfathi**]: https://github.com/soroushfathi
 [**@adnathanail**]: https://github.com/adnathanail
+[**@morrishuismans**]: https://github.com/morrishuismans
 
 <!-- General links -->